Update run.sh to check for ANDROID_HOME

+++ test/run.sh
@@ -1,11 +1,16 @@
 #!/bin/sh
 cd $(dirname $0)
 
+if [ ! -d "$ANDROID_HOME" ]; then
+  echo "ANDROID_HOME is not available"
+  exit
+fi
+
 cd ../complete
 mvn clean package
 ret=$?
 if [ $ret -ne 0 ]; then
-exit $ret
+  exit $ret
 fi
 rm -rf target
 
@@ -13,7 +18,7 @@ cd ../initial
 mvn clean package
 ret=$?
 if [ $ret -ne 0 ]; then
-exit $ret
+  exit $ret
 fi
 rm -rf target
